fix: align solana helper errors
